==> notebooks/processing/0_get_core_data.ipynb <==
import pandas as pd

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
#pd.set_option('display.max_colwidth', None)

df = pd.read_excel("/Users/Peter/work/RADx-rad/RADx-rad Funding Numbers and PIs_2024-03-04_pr.xlsx")
df["phs"] = df["phs"] + ".v1.p1"

df.query("phs == 'phs002574.v1.p1'")

df

dbgap = df[["phs"]].dropna()
dbgap.rename(columns={"phs": "dbgapAccession"}, inplace=True)
dbgap["dbgapAccession"] = "dbgap:" + dbgap["dbgapAccession"]
dbgap.to_csv("../data/Dataset_list.csv", index=False)

projects = pd.read_csv(
        "https://docs.google.com/spreadsheets/d/1ZYUmUJEED3X6Mdo_sTMzOtIRW8IjsqWkAeLBKWdb31k/export?format=csv",
        keep_default_na=False,
       )
# duplicate, this phs number does not exist
projects.query("Accession != 'phs002574.v1.p1'", inplace=True)

projects = projects.query("updated_radx_project == 'RADx-rad'").copy()

projects.query("Accession == 'phs002574.v1.p1'")

projects.head(1000)

projects = projects[["Accession", "sub_project"]].copy()

## Prepare Grant Data

grants = df.merge(projects, left_on="phs", right_on="Accession", how="right")

grants

grants.rename(columns={"DCC": "researchInitiative", "NIH RePORTER Grant Number ": "projectNum", "sub_project": "subProject"}, inplace=True)

grants = grants[["researchInitiative", "projectNum", "subProject"]].copy()

grants["projectNum"] = grants["projectNum"].str.split(";")

grants = grants.explode("projectNum")
grants["projectNum"] = grants["projectNum"].str.strip()

# Make corrections

# remove invalid project number (Douglas Bell)
grants.query("projectNum != 'Z01-ES100475-07'", inplace=True)


grants.at[46, "projectNum"] = "1U01DA053976-01"
grants.at[46, "researchInitiative"] = "RADx-rad"
# Add data for RADx-rad DCC
grants.at[49, "projectNum"] = "1U24LM013755-01"
grants.at[49, "researchInitiative"] = "RADx-rad"

grants

grants.sort_values(by="subProject", inplace=True)
grants

grants.to_csv("../data/Grant_list.csv", index=False)
