==> notebooks/processing/1_get_core_project_info.ipynb <==
# get_core_project_data
This notebook retrieves up-to-date and manually cleaned up RADx core data obtained from the dbGaP database dump received on 2023-08-24.
All other notebooks in this repository depend on the data file generated by this notebook: "../data/radx-projects.csv"

import os
import grant_query
import publication_query
import pandas as pd
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

DERIVED_DATA_PATH = "../derived_data"
os.makedirs(DERIVED_DATA_PATH, exist_ok=True)

projects = pd.read_csv(
        "https://docs.google.com/spreadsheets/d/1ZYUmUJEED3X6Mdo_sTMzOtIRW8IjsqWkAeLBKWdb31k/export?format=csv",
        keep_default_na=False,
       )

projects["project_serial_num"] = projects["core_project_num"].apply(lambda x: grant_query.extract_project_serial_num(x))

## Rename, select, and reorder columns

projects_map = {"updated_radx_project": "researchInitiative", "Accession updated": "dbgapAccession",
                "Grant ID updated": "projectNum", "core_project_num": "coreProjectNum", 
                "project_serial_num": "projectSerialNum", "PI updated": "studyInvestigator", "Study Name": "studyTitle", "sub_project": "subProject"}
projects = publication_query.rename_and_reorder_columns(projects, projects_map)

projects.head()

# remove dbgap_accession duplicates
projects.query("dbgapAccession != 'phs002574.v1.p1'", inplace=True)
projects.query("dbgapAccession != 'phs002516.v1.p1'", inplace=True)

projects.to_csv(os.path.join(DERIVED_DATA_PATH, "radx-projects.csv"), index=False)

print(projects.shape[0])
projects.head()

radx_rad = projects.query("researchInitiative == 'RADx-rad'").copy()
radx_rad["PHS Number"] = radx_rad["dbgapAccession"].str.removesuffix(".v1.p1")
radx_rad = radx_rad[["PHS Number", "studyInvestigator", "projectNum"]].copy()
print(radx_rad.shape[0])
radx_rad

# RADx Data Submission Project Tracker - RADx-rad sheet
#doc_id = "1zBXU_bSPZBJuUFyzmmPgGTqSR0Xdk3IrQOGZ8tv5k-M"
#grid_id = "852835732"

# 2024-03-04:
#doc_id = "1vgCyfi-9Pvlm5-yJdILxws6b1FRtnegO"
#grid_id = "1628495610"
#df = pd.read_csv(f"https://docs.google.com/spreadsheets/d/{doc_id}/export?format=csv&gid={grid_id}", keep_default_na=False)
df = pd.read_excel("/Users/Peter/work/RADx-rad/RADx-rad Funding Numbers and PIs_2024-03-04_pr.xlsx")
print(df.shape[0])
df.head()

df_merge = df.merge(radx_rad, left_on="phs", right_on="PHS Number", how="outer")
df_merge.fillna("", inplace=True)
print(df_merge.shape[0])
df_merge
