==> notebooks/processing/2_get_grant_info.ipynb <==
# get_grant_info
This notebook creates grant-related node and relationship files for the RADx-KG:
* Grant.csv
* FundingOpportunity.csv
* FundingOpportunity-PROVIDES-Grant.csv

import os
import grant_query
import pandas as pd
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
#pd.set_option('display.max_colwidth', None)

KG_PATH = "../kg/data"
DERIVED_DATA_PATH = "../derived_data"

START_FISCAL_YEAR = 2021 # starts Oct. 1, 2020

projects = pd.read_csv(os.path.join(DERIVED_DATA_PATH, "radx-projects.csv"))
projects.query("researchInitiative == 'RADx-rad'", inplace=True)

print(f"Number of projects {projects.shape[0]}")
projects.head()

project_num = list(projects["coreProjectNum"].unique())
print("Number of unique core project numbers:", len(project_num))

### Create Grant nodes

grants = grant_query.get_projects(project_num)
grants["applId"] = grants["applId"].astype(str)
#grants = grant_query.get_projects(project_num)
grants["fiscalYear"] = grants["fiscalYear"].astype(int)

grants = grants[(grants["fiscalYear"] >= START_FISCAL_YEAR) | grants["fiscalYear"].isna()]

grants.sort_values(by="fiscalYear", inplace=True)
grants.drop_duplicates(subset=["coreProjectNum", "fiscalYear"], keep='last', inplace=True)

grants["agency"].unique()

print("Number of grants:", grants.shape[0])

grants["fiscalYear"].value_counts()

grant_nodes = grants.merge(projects[["coreProjectNum", "researchInitiative", "subProject"]], on="coreProjectNum", how="right")
print("Number of grant nodes:", grant_nodes.shape[0])

grant_node_map = {"coreProjectNum": "id", "projectTitle": "name", "abstract": "abstract", 
                  "narrative": "narrative", "fundingMechanism": "fundingMechanism",
                  "awardCode": "awardCode", "researchInitiative": "researchInitiative", "subProject": "subProject"}

grant_nodes = grant_nodes[grant_node_map.keys()].copy()
grant_nodes.rename(columns=grant_node_map, inplace=True)
grant_nodes.drop_duplicates(inplace=True)
grant_nodes.fillna("", inplace=True)

# Remove common headers from abstract and narrative text
grant_nodes["abstract"] = grant_nodes["abstract"].apply(grant_query.remove_prefix, prefix="project")
grant_nodes["abstract"] = grant_nodes["abstract"].apply(grant_query.remove_prefix, prefix="abstract")
grant_nodes["abstract"] = grant_nodes["abstract"].apply(grant_query.remove_prefix, prefix="summary")
grant_nodes["narrative"] = grant_nodes["narrative"].apply(grant_query.remove_prefix, prefix="project")
grant_nodes["narrative"] = grant_nodes["narrative"].apply(grant_query.remove_prefix, prefix="narrative")
grant_nodes["narrative"] = grant_nodes["narrative"].apply(grant_query.remove_prefix, prefix="public health narrative")
grant_nodes["narrative"] = grant_nodes["narrative"].apply(grant_query.remove_prefix, prefix="relevance to public health statement")

# TODO check where name of grant varies for the same grant id
grant_nodes.drop_duplicates(subset="id", inplace=True)
print(grant_nodes.shape[0])
grant_nodes

grant_nodes.to_csv(os.path.join(KG_PATH, "nodes", "Grant.csv"), index=False)
print("Number of grants:", grant_nodes.shape[0])
grant_nodes

print(f"Number of grant nodes: {grant_nodes.shape[0]}")

projects.head()

project_num = set(projects["projectSerialNum"].unique())
grant_num = set(grants["projectSerialNum"].unique())
print("Matching grants:", len(grant_num))

missing = list(project_num - grant_num)
print("Missing grants:", len(missing))

projects[projects["projectSerialNum"].isin(missing)]

### Create FundingOpportunity-PROVIDES-Grant relationships

grant_funding_opportunities = grants[["coreProjectNum", "opportunityNumber"]].copy()
grant_funding_opportunities.dropna(inplace=True)
grant_funding_opportunities.drop_duplicates(inplace=True)

gfo_map = {"opportunityNumber": "from", "coreProjectNum": "to"}

gfo = grants[gfo_map.keys()].copy()
gfo.rename(columns=gfo_map, inplace=True)
gfo.drop_duplicates(inplace=True)
gfo = gfo[gfo["from"] != ""]
gfo

gfo.to_csv(os.path.join(KG_PATH, "relationships", "FundingOpportunity-PROVIDES-Grant.csv"), index=False)
print("Number of funding opportunities:", gfo.shape[0])
gfo.head()

funding_opportunities = gfo["from"].unique().tolist()

fo_details = grant_query.search_funding_opportunities(funding_opportunities)

fo_details.head()

### Create FundingOpportunity Nodes

fo_map = {"number": "id", "title": "name"}
fo = fo_details[fo_map.keys()].copy()
fo.rename(columns=fo_map, inplace=True)
fo.drop_duplicates(inplace=True)

fo["url"] = fo["id"].apply(grant_query.add_funding_opportunity_url)

fo.to_csv(os.path.join(KG_PATH, "nodes", "FundingOpportunity.csv"), index=False)
print("Number of funding opportunities:", fo.shape[0])
fo
