==> notebooks/processing/3_get_citations.ipynb <==
# Get Author info from papers
https://api.semanticscholar.org/api-docs/graph#tag/Paper-Data/operation/get_graph_get_paper_authors

https://api.semanticscholar.org/api-docs/graph#tag/Author-Data/operation/post_graph_get_authors

See also about PubMed unique author id:
https://pubmed.ncbi.nlm.nih.gov/34180522/
https://pubmed.ncbi.nlm.nih.gov/28960025/
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5639053/
https://github.com/Daniel-Mietchen/ideas/issues/1260
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5530597/

import os
import requests
import json
import pandas as pd
import publication_query
import utils

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
#pd.set_option('display.max_colwidth', None)

KG_PATH = "../kg/data"
DERIVED_DATA_PATH = "../derived_data"

START_PUBLICATION_YEAR = 2021 # Grants started at the end of 2020.

## Get raw primary Publication data
These publications result from a grant number query in PubMed.

raw_publications = pd.read_csv("https://raw.githubusercontent.com/christian-horgan/pubmed-search/main/data/radx-rad/article_data_current.csv", dtype=str, keep_default_na=False)

non_pubmed_publications = pd.read_csv("../data/Publications_non_pubmed.csv", dtype=str, keep_default_na=False)
non_pubmed_publications.head()

raw_publications = pd.concat([raw_publications, non_pubmed_publications])
raw_publications.fillna("", inplace=True)

# Restrict publications to 2021+ since RADx projects started at the end of 2020.
raw_publications["publication_year"] = raw_publications["publication_year"].astype(int)
raw_publications.query(f"publication_year >= {START_PUBLICATION_YEAR}", inplace=True)
# DOI is required as the primary key
raw_publications.dropna(subset="doi", inplace=True)
# convert property name to Neo4j convention
raw_publications.rename(columns={"pm_id": "pmId", "pmc_id": "pmcId", "project_serial_num": "projectSerialNum"}, inplace=True)

raw_publications.head()

raw_publications.tail()

## Get author information for the primary citations

# Query the Semantic Scholar API to get author information for a list of DOIs
raw_dois = list(raw_publications["doi"].unique())
authors = publication_query.get_author_ids(raw_dois)
authors.head()

# authors.query("doi == 'doi:10.1097/PCC.0000000000002976'") # Authors with title in names
authors.query("doi == 'doi:10.1109/ICHI54592.2022.00070'")

publication_query.expand_name_column(authors, "name")
authors.rename(columns={"name": "author"}, inplace=True)
authors.rename(columns={"externalIds.ORCID": "orcid"}, inplace=True)
authors = raw_publications.merge(authors, on="doi")

authors.drop_duplicates(inplace=True)
authors.fillna("", inplace=True)
authors.replace("nan", "", inplace=True)
authors.rename(columns={"name": "author"}, inplace=True)

## Match publication authors with RADx-rad investigators
In this step, we eliminate non-RADx publications.

radx_investigators = pd.read_csv(os.path.join(DERIVED_DATA_PATH, "radx_investigators.csv"))

radx_investigators.head()

#### Fuzzy merge of publication authors with RADx investigators

# The threshold was manually adjusted by checking for proper matches in the table below. The lowest threshold is 0.905882 for author Vanegas D.
radx_authors = utils.fuzzy_merge(authors[["author", "authorId", "aliases", "orcid", "projectSerialNum", "doi"]], 
                                    radx_investigators[["name", "projectSerialNum"]], 
                                    left_fuzzy_on="author", right_fuzzy_on="name", 
                                    left_on="projectSerialNum", right_on="projectSerialNum", 
                                    how="inner", threshold=0.905)

print("Matched authors       :", radx_authors.shape[0])
print("Matched unique authors:", radx_authors["author"].nunique())
print("Number of unique DOIs :", radx_authors["doi"].nunique())
radx_authors.head()

radx_authors.query("doi == 'doi:10.1109/ICHI54592.2022.00070'")

#### If there are multiple name matches for a publication, use the match with the highest score.

# sort descending so the match with the highest score comes first.
radx_authors.sort_values(by=["score", "doi", "match"], ascending=False, inplace=True)
# keep only the first instance for each match, thus eliminating the lower-scoring matches.
radx_authors.drop_duplicates(subset=["doi", "match"], inplace=True)
print("Matched authors       :", radx_authors.shape[0])
print("Matched unique authors:", radx_authors["author"].nunique())
print("Number of unique DOIs :", radx_authors["doi"].nunique())
radx_authors.head()

#### Filter publications (RADx investigators only)

primary_dois = list(radx_authors["doi"].unique())
primary_publications = raw_publications[raw_publications["doi"].isin(set(primary_dois))].copy()
primary_publications.fillna("", inplace=True)
primary_publications.head()

## Create primary ```Publication``` node file for KG

primary_publications["id"] = primary_publications["doi"]
primary_publications["url"] = "https://doi.org/" + primary_publications["doi"]
primary_publications["type"] = "primary"
primary_publications.drop_duplicates("id", inplace=True)

publication_map = {"id": "id", "title": "name", "abstract": "abstract", "journal_name": "journal", "publication_year": "year", "type": "type", "doi": "doi", "pmId": "pmId", "pmcId": "pmcId", "url": "url"}
primary_publications_kg = utils.rename_and_reorder_columns(primary_publications, publication_map)

primary_publications_kg.to_csv(os.path.join(KG_PATH, "nodes", "Publication_primary.csv"), index=False)

### Create primary publication Author nodes

primary_authors_all = authors[authors["doi"].isin(set(primary_dois))].copy()
primary_authors_all["fullName"] = primary_authors_all["fullName"].str.replace(".", "")
primary_authors_all.head()

#primary_authors[["authorId", "author", "fullName", "firstName", "middleName", "lastName", "doi"]].to_csv(os.path.join(DERIVED_DATA_PATH, "primary_authors_all.csv"), index=False)

## Create ```Publication-CITES-Publication``` relationship file for KG

citations = publication_query.get_citations(primary_dois)
citations.rename(columns={"doiCite": "from", "doi": "to"}, inplace=True)
# TODO where does the following inconsistency originate?
# WARNING: Error messages from data import:
# doi:10.47464/METROCIENCIA/VOL29/3/2021/5-10 (Publication-ID)-[CITES]->doi:10.1002/art.41616 (Publication-ID) referring to missing node doi:10.47464/METROCIENCIA/VOL29/3/2021/5-10
citations = citations[(citations["from"] != "doi:10.47464/METROCIENCIA/VOL29/3/2021/5-10") & (citations["to"] != "doi:10.1002/art.41616")]
citations.to_csv(os.path.join(KG_PATH, "relationships", "Publication-CITES-Publication.csv"), index=False)

## Get Secondary publications
Secondary publications are publications that cite the primary publication

secondary_dois = list(citations["from"].unique())
# remove any primary DOIs, e.g., a primary publication cites another primary publication.
secondary_dois = list(set(secondary_dois) - set(primary_dois))
secondary_publications = publication_query.get_publication_info(secondary_dois)

secondary_publications["id"] = secondary_publications["doi"]
secondary_publications["url"] = "https://doi.org/" + secondary_publications["doi"]
secondary_publications["type"] = "secondary"
secondary_publications.drop_duplicates("id", inplace=True)

citation_map = {"id": "id", "title": "name", "abstract": "abstract", "journal": "journal", "year": "year", "type": "type", "doi": "doi", "pmId": "pmId", "pmcId": "pmcId", "url": "url"}
secondary_publications = utils.rename_and_reorder_columns(secondary_publications, citation_map)

print("Number of secondary publications:", secondary_publications.shape[0])
secondary_publications.head()

secondary_publications.to_csv(os.path.join(KG_PATH, "nodes", "Publication_secondary.csv"), index=False)

### Create list of primary Authors

radx_authors.head()

primary_authors = radx_authors[["authorId", "author", "aliases", "projectSerialNum", "doi"]].copy()

primary_authors.drop_duplicates(inplace=True)
primary_authors.dropna(inplace=True)
primary_authors.sort_values("author", inplace=True)
print(primary_authors.shape[0])
primary_authors.head()

primary_authors.to_csv(os.path.join("../derived_data/", "primary_authors.csv"), index=False)

## Create a list of primary authors who are not PIs or Study Investigators

primary_authors_list = list(primary_authors["authorId"].unique())

primary_authors_all_list = list(primary_authors_all["authorId"].unique())
primary_authors_other_list = set(primary_authors_all_list) - set(primary_authors_list)

primary_authors_other = primary_authors_all[primary_authors_all["authorId"].isin(primary_authors_other_list)]
primary_authors_other = primary_authors_other[["authorId", "author", "fullName", "firstName", "middleName", "lastName", "doi"]].copy()
primary_authors_other.drop_duplicates(inplace=True)
primary_authors_other.dropna(inplace=True)
primary_authors_other.sort_values("author", inplace=True)

primary_authors_other.to_csv(os.path.join("../derived_data/", "primary_authors_other.csv"), index=False)
print(f"Number of other primary paper authors: {primary_authors_other.shape[0]}")
primary_authors_other.head()

### Create list of secondary authors

# secondary_authors = publication_query.get_author_ids(secondary_dois)
# author_map = {"authorId": "authorId", "title": "name", "abstract": "abstract", "journal": "journal", "year": "year", "type": "type", "doi": "doi", "pmId": "pmId", "pmcId": "pmcId", "url": "url"}
# secondary_authors = utils.rename_and_reorder_columns(secondary_authors, author_map)
# secondary_authors.rename(columns={"name": "author"})
# secondary_authors.head()
#secondary_authors = authors[["authorId", "author", "aliases", "orcid", "projectSerialNum"]].copy()

#secondary_authors.to_csv(os.path.join("../derived_data/", "secondary_authors.csv"), index=False)
