==> notebooks/processing/5_process_publications.ipynb <==
import os
import pandas as pd
import grant_query
import publication_query
import utils
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
#pd.set_option('display.max_colwidth', None)

KG_PATH = "../kg/data"
DERIVED_DATA_PATH = "../derived_data"

primary_authors = pd.read_csv(os.path.join(DERIVED_DATA_PATH, "primary_authors.csv"), dtype=str, keep_default_na=False)
primary_authors.head()

id_to_author_id = pd.read_csv(os.path.join(DERIVED_DATA_PATH, "id_to_author_id.csv"), dtype=str, keep_default_na=False)

id_to_author_id.head()

primary_authors = primary_authors.merge(id_to_author_id, on="authorId", how="left")
primary_authors.fillna("")
primary_authors.head()

## Create Researcher-AUTHORED-Publication nodes

author_map = {"id": "from", "doi": "to"}
authored = utils.rename_and_reorder_columns(primary_authors, author_map)
authored = authored[(authored["from"] != "") & (authored["to"] != "")]
authored.drop_duplicates(inplace=True)
authored = authored[authored["from"] != "s2authorid:None"]
authored.to_csv(os.path.join(KG_PATH, "relationships", "Researcher-AUTHORED-Publication_investigators.csv"), index=False)
print(f"Number of other Researcher nodes: {authored.shape[0]}")
authored.head()

primary_authors_other = pd.read_csv(os.path.join("../derived_data/", "primary_authors_other.csv"), dtype=str, keep_default_na=False)
primary_authors_other["id"] = "s2authorid:" + primary_authors_other["authorId"]
authored_other = utils.rename_and_reorder_columns(primary_authors_other, author_map)
authored_other.drop_duplicates(inplace=True)
authored_other = authored_other[authored_other["from"] != "s2authorid:None"]
authored_other.to_csv(os.path.join(KG_PATH, "relationships", "Researcher-AUTHORED-Publication_other.csv"), index=False)
print(f"Number of other Researcher nodes: {authored_other.shape[0]}")
authored_other.head()
