==> notebooks/processing/6_get_coordination_center_info.ipynb <==
# get_research_initiative_info
This notebook creates coordination center-related node and relationship files for the RADx-KG:
* ResearchInitiative.csv
* Grant-FUNDED-ResearchInitiative.csv
* ResearchInitiative-SUPPORTED-Grant.csv

import os
import shutil
import pandas as pd
pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", None)
pd.set_option("display.max_colwidth", None)

KG_PATH = "../kg/data"
DERIVED_DATA_PATH = "../derived_data"

### Create ResearchInitiative Nodes
This information is maintained manually. It's copied fro the data directory.

shutil.copy("../data/ResearchInitiative.csv", os.path.join(KG_PATH, "nodes"))

### Create Grant-FUNDED-ResearchInitiative relationships
This information is maintained manually. It's copied fro the data directory.

shutil.copy("../data/Grant-FUNDED-ResearchInitiative.csv", os.path.join(KG_PATH, "relationships"))

projects = pd.read_csv(os.path.join(DERIVED_DATA_PATH, "radx-projects.csv"))
projects.query("researchInitiative == 'RADx-rad'", inplace=True)
#projects.query("radx_project == "RADx-UP"", inplace=True)

print(f"Number of projects {projects.shape[0]}")
projects.head()

### Create ResearchInitiative-SUPPORTED-Grant

supported_map = {"researchInitiative": "from", "coreProjectNum": "to"}


supported = projects[supported_map.keys()].copy()
supported.rename(columns=supported_map, inplace=True)
supported.dropna(inplace=True)
supported.drop_duplicates(inplace=True)

# Exclude grants that support the Coordination Centers (RADx-rad, RADx-UP, RADx-HUB)
excluded_projects = ["U24LM013755", "U24MD016258", "OT2DB000009"]
supported = supported[~supported["to"].isin(excluded_projects)]

supported.to_csv(os.path.join(KG_PATH, "relationships", "ResearchInitiative-SUPPORTED-Grant.csv"), index=False)

supported.head()
