==> notebooks/processing/7_get_patent_info.ipynb <==
# get_patent_info
This notebook creates dataset-related node and relationship files for the RADx-KG:
* Patent.csv
* Researcher-IS_INVENTOR_OF-Patent
* Patent-IS_RELATED_TO-Grant.csv

import os
import pandas as pd
import utils
import publication_query
pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", None)
pd.set_option("display.max_colwidth", None)

KG_PATH = "../kg/data"

patent_info = pd.read_csv(
        "https://docs.google.com/spreadsheets/d/1O_IpO8-1IMHXmhpEw6eRU6KPNtBtnJ38GEL-tx9F9Mw/export?format=csv",
        keep_default_na=False,
       )

print(f"Number of projects {patent_info.shape[0]}")
patent_info

patent_inventors = patent_info[["inventors", "patentId"]].copy()
patent_inventors["inventors"] = patent_inventors["inventors"].str.split("|")
patent_inventors = patent_inventors.explode("inventors")
#publication_query.expand_name_column(patent_inventors, "inventors")
#patent_inventors.rename(columns={"name": "inventor"}, inplace=True)
patent_grants = patent_info[["coreProjectNum", "patentId"]].copy()
patent = patent_info[["patentId", "title", "inventors", "filingDate", "grantedDate", "status", "applicants"]].copy()

investigators = pd.read_csv(os.path.join(KG_PATH, "nodes", "Researcher_investigators.csv"), dtype=str, keep_default_na=False)
other = pd.read_csv(os.path.join(KG_PATH, "nodes", "Researcher_primary_coauthors.csv"), dtype=str, keep_default_na=False)
researchers = pd.concat([investigators, other])
researchers.head()

investigators.head()

research_inventors = utils.fuzzy_merge(patent_inventors, researchers, left_fuzzy_on="inventors", right_fuzzy_on="fullName", how="left", threshold=0.95)

research_inventors.head()

research_inventors_no_match = research_inventors.query("id == ''").copy()
research_inventors_no_match.head()

researchers["firstLastName"] = researchers["firstName"] + " " + researchers["lastName"]
researchers.head()

publication_query.expand_name_column(research_inventors_no_match, "inventors")
research_inventors_no_match["inventor"] = research_inventors_no_match["firstName"] + " " + research_inventors_no_match["lastName"]

research_inventors_no_match = research_inventors_no_match[["inventors", "inventor", "patentId"]].copy()

research_inventors_no_match = utils.fuzzy_merge(research_inventors_no_match, researchers, left_fuzzy_on="inventor", right_fuzzy_on="firstLastName", how="left", threshold=0.96)

research_inventors_no_match.head()

## Create Researcher-IS_INVENTOR-Patent relationships

research_inventors = research_inventors[["id", "patentId"]].copy()
research_inventors.query("id != ''", inplace=True)

research_inventors_no_match = research_inventors_no_match[["id", "patentId"]].copy()
research_inventors_no_match.query("id != ''", inplace=True)

research_inventors_matched = pd.concat([research_inventors, research_inventors_no_match])

research_inventors_matched.rename(columns={"id": "from", "patentId": "to"}, inplace=True)
research_inventors_matched.to_csv(os.path.join(KG_PATH, "relationships", "Researcher-IS_INVENTOR-Patent.csv"), index=False)
research_inventors_matched.head()                          

## Create Patent nodes

patent_map = {"patentId": "id", "title": "name", "inventors": "inventors", "filingDate": "filingDate",	"grantedDate": "grantedDate", "status": "status", "applicants": "applicants", "patentUrl": "url"}
patent = utils.rename_and_reorder_columns(patent_info, patent_map)

patent.to_csv(os.path.join(KG_PATH, "nodes", "Patent.csv"), index=False)
patent.head()
