==> notebooks/processing/8_get_dataset_usage.ipynb <==
# get_dateset_usage
This notebook creates event-related node and relationship files for the RADx-KG:

Event.csv
Event-RELATED_TO-CoordinationCenter.csv

import os
import shutil
import pandas as pd
import hashlib
import utils
import publication_query
pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", None)
pd.set_option("display.max_colwidth", None)

KG_PATH = "../kg/data"

usage = pd.read_csv("../data/Dataset_usage.csv")
usage.head()

author_map = {"Rose, Peter": "s2authorid:2184494479", "Hong, Na": "s2authorid:2072483999", "Miguez, Maria-Jose": "s2authorid:40344211", 
              "Anwar, Mohd Mozharul": "s2authorid:143740901", "Solo-Gabriele, Helena": "profileid:11165737", "Ciofani, Danielle": "s2authorid:2169822026",
              "Cho, Hyunghoon": "s2authorid:1792468", "Quandelacy, Talia": "s2authorid:1397184885"}

author_map.get("Rose, Peter")

usage["Requestor"] = usage["Requestor"].str.strip()
usage["from"] = usage["Requestor"].apply(lambda x: author_map.get(x))
usage["to"] = "dbgap:" + usage["accession"]
usage = usage[["from", "to"]].copy()

usage.to_csv(os.path.join(KG_PATH, "relationships", "Researcher-USED-Dataset.csv"), index=False)
print(f"Number of Researcher-USED-Dataset relationships: {usage.shape[0]}")
usage
