==> notebooks/processing/8_get_event_info.ipynb <==
# get_event_info
This notebook creates event-related node and relationship files for the RADx-KG:

Event.csv
Event-RELATED_TO-CoordinationCenter.csv

import os
import shutil
import pandas as pd
import hashlib
import utils
import publication_query
pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", None)
pd.set_option("display.max_colwidth", None)

KG_PATH = "../kg/data"

# Event spreadsheet
doc_id = "1ZPvaKqHYIusiIROpjImlXLC2tqZFnkwb"
grid_id = "1054557813"
presentations = pd.read_csv(f"https://docs.google.com/spreadsheets/d/{doc_id}/export?format=csv&gid={grid_id}", dtype=str, keep_default_na=False)
# assign unique ID to each presentation
presentations["presentationId"] = presentations["eventId"] + presentations["title"]
presentations["presentationId"] = presentations["presentationId"].apply(lambda x: hashlib.md5(x.encode()).hexdigest())
presentations.head()
print(presentations.shape[0])
presentations.head()

## Create Presentation-PRESENTED_AT-Event relationships

presentations_map = {"presentationId": "from", "eventId": "to"}
presented = utils.rename_and_reorder_columns(presentations, presentations_map)
presented.to_csv(os.path.join(KG_PATH, "relationships", "Presentation-PRESENTED_AT-Event.csv"), index=False)
print(f"Number of Presentation nodes: {presented.shape[0]}")
presented.head()

#presentations["id"] = presentations["title"].apply(lambda x: hashlib.md5(x.encode()).hexdigest())
#presentations.head()

presenters = presentations[["eventId", "presenters", "presentationId"]].copy()
presenters["presenters"] = presenters["presenters"].str.split("|")
presenters = presenters.explode("presenters")
publication_query.expand_name_column(presenters, "presenters")
presenters["presenters"] = presenters["name"]
presenters.drop(columns="name", inplace=True)
presenters.head()

investigators = pd.read_csv(os.path.join(KG_PATH, "nodes", "Researcher_investigators.csv"), dtype=str, keep_default_na=False)
other = pd.read_csv(os.path.join(KG_PATH, "nodes", "Researcher_primary_coauthors.csv"), dtype=str, keep_default_na=False)
researchers = pd.concat([investigators, other])
researchers.head()

presenters_matched = utils.fuzzy_merge(presenters, researchers, left_fuzzy_on="presenters", right_fuzzy_on="name", how="inner", threshold=0.95)
presenters_matched.head()

## Create Researcher-PRESENTED-Presentation relationships

presented = presenters_matched[["id", "presentationId"]].copy()
presenters_map = {"id": "from", "presentationId": "to"}
presented = utils.rename_and_reorder_columns(presented, presenters_map)
presented.drop_duplicates(inplace=True)
presented = presented[(presented["from"] != "") & (presented["to"] != "")]
presented.to_csv(os.path.join(KG_PATH, "relationships", "Researcher-PRESENTED-Presentation.csv"), index=False)
print(f"Number of other Researcher nodes: {presented.shape[0]}")
presented.head()

## Create Presentation Nodes

presentations_map = {"presentationId": "id", "title": "name", "presenters": "presenters", "presentationUrl": "presentationUrl",	"videoUrl": "videoUrl"}
presentations = utils.rename_and_reorder_columns(presentations, presentations_map)
presentations.drop_duplicates(inplace=True)
presentations.to_csv(os.path.join(KG_PATH, "nodes", "Presentation.csv"), index=False)
print(f"Number of Presentation nodes: {presentations.shape[0]}")
presentations.head()

## Create Event nodes
Event info is maintained manually in the data directory

doc_id = "1ZPvaKqHYIusiIROpjImlXLC2tqZFnkwb"
grid_id = "77214925"
events = pd.read_csv(f"https://docs.google.com/spreadsheets/d/{doc_id}/export?format=csv&gid={grid_id}", usecols=["id", "name", "eventType", "eventUrl", "startDate", "endDate", "city", "state", "country"], dtype=str, keep_default_na=False)
events.to_csv(os.path.join(KG_PATH, "nodes", "Event.csv"), index=False)
print(f"Number of Event nodes: {events.shape[0]}")
events.head()

#shutil.copy("../data/Event.csv", os.path.join(KG_PATH, "nodes"))

## Create Event-RELATED_TO-CoordinationCenter.csv

#shutil.copy("../data/Event-RELATED_TO-CoordinationCenter.csv", os.path.join(KG_PATH, "relationships"))
