==> notebooks/processing/8_get_manual_data.ipynb <==
# 8_get_manual_data
import os
import shutil

KG_PATH = "../kg/data"

shutil.copy("../data/Software_manually.csv", os.path.join(KG_PATH, "nodes"))

shutil.copy("../data/Researcher-DEVELOPED-Software.csv", os.path.join(KG_PATH, "relationships"))

shutil.copy("../data/Dataset_manually.csv", os.path.join(KG_PATH, "nodes"))

shutil.copy("../data/Researcher_manually.csv", os.path.join(KG_PATH, "nodes"))

shutil.copy("../data/Researcher-CREATED-Dataset_manually.csv", os.path.join(KG_PATH, "relationships"))

## Add other Researcher-IS_INVESTIGATOR_OF-Grant
Misc. researchers that are maintained manually (e.g., software developers not yet included in the researcher nodes created above.

shutil.copy("../data/Researcher-IS_INVESTIGATOR_OF-Grant_manually.csv", os.path.join(KG_PATH, "relationships"))
