==> notebooks/processing/9_add_embeddings.ipynb <==
# Add Embeddings for Publications
[work in progress]

import os
import shutil
import pandas as pd
from typing import Dict, List
import json
import requests

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
#pd.set_option('display.max_colwidth', None)

KG_PATH = "../kg/data"

## Calculate Embeddings for Publications
A joint embedding of the publication title and abstract is calculated using the [SPECTER](https://github.com/allenai/specter) [public API](https://github.com/allenai/paper-embedding-public-apis).

data_file_primary = os.path.join(KG_PATH, "nodes", "Publication_primary.csv")

publications_primary = pd.read_csv(data_file_primary, dtype=str)
publications_primary.head()

# create a JSON object with the fields required by the SPECTER API
publications_part = publications_primary[["id", "name", "abstract"]].copy()
publications_part.rename(columns={"id": "paper_id", "name": "title"}, inplace=True)
publications_dict = json.loads(publications_part.to_json(orient="records"))

URL = "https://model-apis.semanticscholar.org/specter/v1/invoke"
MAX_BATCH_SIZE = 16

def chunks(lst, chunk_size=MAX_BATCH_SIZE):
    """Splits a longer list to respect batch size"""
    for i in range(0, len(lst), chunk_size):
        yield lst[i : i + chunk_size]

def embed(papers):
    embeddings_by_paper_id: Dict[str, List[float]] = {}

    for chunk in chunks(papers):
        # Allow Python requests to convert the data above to JSON
        response = requests.post(URL, json=chunk)

        if response.status_code != 200:
            raise RuntimeError("Sorry, something went wrong, please try later!")

        for paper in response.json()["preds"]:
            embeddings_by_paper_id[paper["paper_id"]] = paper["embedding"]

    return embeddings_by_paper_id

# reformat embeddings into a vertical bar separated array

embeddings = pd.DataFrame(embed(publications_dict)).transpose()
embeddings["embedding"] = [[e for e in row if isinstance(e, float) ] for row in embeddings.values.tolist()]
embeddings["embedding"] = embeddings["embedding"].astype(str)
# remove brackets from Python list
embeddings["embedding"] = embeddings["embedding"].str.replace("[", "")
embeddings["embedding"] = embeddings["embedding"].str.replace("]", "")
# replace comma with vertical bar delimiter
embeddings["embedding"] = embeddings["embedding"].str.replace(", ", "|")
embeddings = embeddings.reset_index(names="id")
embeddings = embeddings[["id", "embedding"]].copy()

embeddings.head()

publications_primary.head()

# append embeddings to publication nodes
publications_primary = publications_primary.merge(embeddings, on="id")

publications_primary.head()

# save publications with embedding added.
publications_primary.to_csv(data_file_primary, index=False)

# don't calculate embeddings for secondary publications for now
data_file_secondary = os.path.join(KG_PATH, "nodes", "Publication_secondary.csv")
publications_secondary = pd.read_csv(data_file_secondary, dtype=str)
publications_secondary["embedding"] = ""
# save publications with embedding added.
publications_secondary.to_csv(data_file_secondary, index=False)
